Tests align with moved expand_table_fields
