+++ dbt_dry_run/test/test_columns_metadata.py
@@ -2,11 +2,16 @@ from typing import Dict, List
 
 import pytest
 
-from dbt_dry_run.columns_metadata import REPEATED_SUFFIX, map_columns_to_table
+from dbt_dry_run.columns_metadata import (
+    REPEATED_SUFFIX,
+    expand_table_fields,
+    map_columns_to_table,
+)
 from dbt_dry_run.exception import InvalidColumnSpecification, UnknownDataTypeException
 from dbt_dry_run.literals import enable_test_example_values
 from dbt_dry_run.models import BigQueryFieldMode, BigQueryFieldType, Table, TableField
 from dbt_dry_run.models.manifest import ManifestColumn
+from dbt_dry_run.test.utils import field_with_name
 
 enable_test_example_values(True)
 
@@ -26,6 +31,48 @@ def assert_columns_result_in_table(
 
 def field_type_as_repeated(field_type: BigQueryFieldType) -> str:
     return field_type + REPEATED_SUFFIX
+
+
+def test_expand_table_fields_with_column_names_with_no_nesting() -> None:
+    table = Table(fields=[field_with_name("a"), field_with_name("b")])
+
+    expected = {"a", "b"}
+    actual = expand_table_fields(table)
+    assert actual == expected
+
+
+def test_expand_table_fields_with_struct() -> None:
+    table = Table(
+        fields=[
+            field_with_name("a"),
+            field_with_name(
+                "struct",
+                fields=[field_with_name("struct_1"), field_with_name("struct_2")],
+            ),
+        ]
+    )
+
+    expected = {"a", "struct", "struct.struct_1", "struct.struct_2"}
+    actual = expand_table_fields(table)
+    assert actual == expected
+
+
+def test_expand_table_fields_with_nested_struct() -> None:
+    table = Table(
+        fields=[
+            field_with_name("a"),
+            field_with_name(
+                "struct",
+                fields=[
+                    field_with_name("struct_1", fields=[field_with_name("struct_1_1")])
+                ],
+            ),
+        ]
+    )
+
+    expected = {"a", "struct", "struct.struct_1", "struct.struct_1.struct_1_1"}
+    actual = expand_table_fields(table)
+    assert actual == expected
 
 
 def test_map_columns_to_table_handles_flat_schema() -> None:
