+++ dbt_dry_run/test/utils.py
@@ -2,6 +2,7 @@ from typing import List, Optional, Union
 
 from pydantic import BaseModel, Field
 
+from dbt_dry_run.models import BigQueryFieldMode, BigQueryFieldType, TableField
 from dbt_dry_run.models.manifest import Node, NodeConfig, NodeDependsOn, NodeMeta
 from dbt_dry_run.scheduler import ManifestScheduler
 
@@ -46,3 +47,12 @@ class SimpleNode(BaseModel):
 
 
 SimpleNode.update_forward_refs()
+
+
+def field_with_name(
+    name: str,
+    type_: BigQueryFieldType = BigQueryFieldType.STRING,
+    mode: BigQueryFieldMode = BigQueryFieldMode.NULLABLE,
+    fields: Optional[List[TableField]] = None,
+) -> TableField:
+    return TableField(name=name, type=type_, mode=mode, fields=fields)
